pyproject.toml
[tool.poetry]
name = "documento-printserver"
version = "0.1.0"
description = "Print server for documento"
authors = ["Jonathan Weth <git@jonathanweth.de>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.6"
python-escpos = "^3.0a8"
python-barcode = "^0.13.1"
requests = "^2.25.1"
dynaconf = "^3.1.4"

[tool.poetry.dev-dependencies]
sphinx = "^3.5"
sphinx-autodoc-typehints = "^1.7"
pytest = "^6.0"
safety = "^1.8.5"
flake8 = "^3.9.0"
flake8-fixme = "^1.1.1"
flake8-mypy = "^17.8.0"
flake8-bandit = "^2.1.2"
flake8-builtins = "^1.4.1"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.0.14"
black = "^19.10b0"
flake8-black = "^0.2.0"
isort = "^5.8.0"
flake8-isort = "^4.0.0"
pytest-cov = "^2.8.1"
pytest-sugar = "^0.9.2"
tox = "^3.23.0"

[tool.black]
line-length = 100

[tool.poetry.scripts]
documento-printserver = "documento_printserver.print_server:print_server"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
